fix: Service engine created
Created services that are called by controllers and external modules
